[fix](test) Make Decimal-from-literal constructions portable on macOS
Building the BE unit tests with the macOS (arm64) toolchain failed to compile:

    error: conversion from 'long' to 'Decimal64' is ambiguous

Decimal64/Decimal128V2/Decimal128V3/Decimal256 are Decimal<T> with several
non-explicit integer constructors (Int32/Int64/Int128/UInt64/...). A bare
integer literal larger than INT_MAX has type 'long'. On Linux LP64 int64_t is
'long', so the literal matches Decimal(Int64) exactly; on macOS int64_t is
'long long', so 'long' converts equally well to several constructor parameter
types and overload resolution is ambiguous.

Wrap these literals in int64_t(...) so they match the Decimal(Int64)
constructor exactly on both platforms. This mirrors the existing convention in
other tests (e.g. jsonb_document_test.cpp, native_reader_writer_test.cpp).

+++ be/test/exec/column_type_convert_test.cpp
@@ -485,7 +485,7 @@ TEST_F(ColumnTypeConverterTest, TestDecimalConversions) {
         src_data.push_back(Decimal256(327673345));
         src_data.push_back(Decimal256(655353345));
         src_data.push_back(Decimal256(655363345));
-        src_data.push_back(Decimal256(3333333333332345));
+        src_data.push_back(Decimal256(int64_t(3333333333332345)));
 
         auto mutable_dst = nullable_dst_type->create_column();
         auto& nullable_col = static_cast<ColumnNullable&>(*mutable_dst);
